% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calIBS_fun.r
\name{calIBS}
\alias{calIBS}
\title{To predict the heterosis of a given breed with other breeds}
\usage{
calIBS(snp_lst, breeds)
}
\arguments{
\item{x}{old score}
}
\value{
new score
}
\description{
To predict the heterosis of a given breed with other breeds
}
\examples{
calIBS(snp_lst,breeds)
}
